+++ buildbot_travis/steps/spawner.py
@@ -2,15 +2,16 @@
 from buildbot.process import buildstep
 from buildbot.process.buildstep import LoggingBuildStep, SUCCESS, FAILURE, EXCEPTION
 from buildbot.process.properties import Properties
+from buildbot.steps.trigger import Trigger
 from twisted.spread import pb
 from twisted.internet import defer
 from twisted.python import log
 import StringIO
 
-from .base import ConfigurableStep
+from .base import ConfigurableStepMixin
 
 
-class TravisTrigger(ConfigurableStep):
+class TravisTrigger(Trigger, ConfigurableStepMixin):
 
     haltOnFailure = True
     flunkOnFailure = True
@@ -20,122 +21,31 @@ class TravisTrigger(ConfigurableStep):
     updateSourceStamp = True
 
     def __init__(self, scheduler, **kwargs):
-        if not "name" in kwargs:
+        if "name" not in kwargs:
             kwargs['name'] = 'trigger'
-        # if not "description" in kwargs:
-        #    kwargs['description'] = kwargs['name']
-        LoggingBuildStep.__init__(self, **kwargs)
-        self.addFactoryArguments(scheduler=scheduler)
-
-        self.scheduler = scheduler
-        self.running = False
-        self.ended = False
-
-    def interrupt(self, reason):
-        if self.running and not self.ended:
-            self.step_status.setText(["interrupted"])
-            return self.end(EXCEPTION)
-
-    def end(self, result):
-        if not self.ended:
-            self.ended = True
-            return self.finished(result)
-
-    def prepareSourcestampListForTrigger(self):
-        if self.sourceStamps:
-            ss_for_trigger = {}
-            for ss in self.sourceStamps:
-                codebase = ss.get('codebase', '')
-                assert codebase not in ss_for_trigger, "codebase specified multiple times"
-                ss_for_trigger[codebase] = ss
-            return ss_for_trigger
-
-        if self.alwaysUseLatest:
-            return {}
-
-        # start with the sourcestamps from current build
-        ss_for_trigger = {}
-        objs_from_build = self.build.getAllSourceStamps()
-        for ss in objs_from_build:
-            ss_for_trigger[ss.codebase] = ss.asDict()
-
-        # overrule revision in sourcestamps with got revision
-        if self.updateSourceStamp:
-            got = self.build.build_status.getAllGotRevisions()
-            for codebase in ss_for_trigger:
-                if codebase in got:
-                    ss_for_trigger[codebase]['revision'] = got[codebase]
-
-        return ss_for_trigger
+        Trigger.__init__(self, waitForFinish=True, schedulerNames=[scheduler], **kwargs)
 
     @defer.inlineCallbacks
-    def start(self):
-        config = yield self.getStepConfig()
+    def run(self):
+        self.config = yield self.getStepConfig()
 
-        ss_for_trigger = self.prepareSourcestampListForTrigger()
-        ss = ss_for_trigger[self.build.builder.name]
+        rv = yield Trigger.run(self)
+        defer.returnValue(rv)
 
-        # Stop the build early if .travis.yml says we should ignore branch
-        if ss.get('branch', None) and not config.can_build_branch(ss['branch']):
-            defer.returnValue(self.end(SUCCESS))
+    def getSchedulers(self):
+        (triggered_schedulers, invalid_schedulers) = Trigger.getSchedulers(self)
+        # baseclass should return one scheduler
+        if invalid_schedulers:
+            return ([], invalid_schedulers)
 
-        # Find the master object
-        master = self.build.builder.botmaster.parent
+        sch = triggered_schedulers[0]
+        triggered_schedulers = []
 
-        # Find the scheduler we are going to use to queue actual builds
-        all_schedulers = self.build.builder.botmaster.parent.allSchedulers()
-        all_schedulers = dict([(sch.name, sch) for sch in all_schedulers])
-        sch = all_schedulers[self.scheduler]
-
-        triggered = []
-
-        self.running = True
-
-        for env in config.matrix:
+        for env in self.config.matrix:
             props_to_set = Properties()
             props_to_set.updateFromProperties(self.build.getProperties())
             props_to_set.update(env["env"], ".travis.yml")
-            props_to_set.setProperty(
-                "spawned_by",  self.build.build_status.number, "Scheduler")
+            props_to_set.setProperty("spawned_by",  self.build.number, "Scheduler")
 
-            triggered.append(
-                sch.trigger(ss_for_trigger, set_props=props_to_set))
-
-        results = yield defer.DeferredList(triggered, consumeErrors=1)
-
-        was_exception = was_failure = False
-        brids = {}
-
-        for was_cb, results in results:
-            if isinstance(results, tuple):
-                results, some_brids = results
-                brids.update(some_brids)
-
-            if not was_cb:
-                was_exception = True
-                log.err(results)
-                continue
-
-            if results == FAILURE:
-                was_failure = True
-
-        if was_exception:
-            result = EXCEPTION
-        elif was_failure:
-            result = FAILURE
-        else:
-            result = SUCCESS
-
-        if brids:
-            brid_to_bn = dict((_brid, _bn) for _bn, _brid in brids.iteritems())
-            res = yield defer.DeferredList([master.db.builds.getBuildsForRequest(br) for br in brids.values()], consumeErrors=1)
-            for was_cb, builddicts in res:
-                if was_cb:
-                    for build in builddicts:
-                        bn = brid_to_bn[build['brid']]
-                        num = build['number']
-
-                        url = master.status.getURLForBuild(bn, num)
-                        self.step_status.addURL("%s #%d" % (bn, num), url)
-
-        defer.returnValue(self.end(result))
+            triggered_schedulers.append((sch, props_to_set))
+        return triggered_schedulers, []
