+++ sample.cfg
@@ -1,8 +1,2 @@
-
-from buildbot_travis import mergeRequests, Loader
-
-c['mergeRequests'] = mergeRequests
-
-l = Loader(c)
-l.load("sample.yml")
-
+from buildbot_travis import TravisConfigurator
+TravisConfigurator(BuildmasterConfig, basedir).fromYaml("sample.yml")
