+++ sample.yml
@@ -1,5 +1,15 @@
+not_important_files:
+    - CHANGES
+    - CHANGES.rst
+    - CHANGES.txt
+    - HISTORY.txt
+    - HISTORY
+    - version.txt
+    - .badger.yml
+
+env:
+    CI: true
 
 projects:
   - name: buildbot_travis
     repository: git://github.com/Jc2k/buildbot_travis
-
